refactor: BetAmount static 메소드 제거 및 패키지 변경
