+++ src/main/java/blackjack/model/participant/Player.java
@@ -1,6 +1,5 @@
 package blackjack.model.participant;
 
-import blackjack.model.BetAmount;
 import blackjack.model.card.CardDeck;
 import blackjack.model.result.Result;
 import blackjack.model.state.FinalState;
