+++ src/test/java/blackjack/model/BetAmountTest.java
@@ -1,5 +1,6 @@
 package blackjack.model;
 
+import blackjack.model.participant.BetAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
