+++ src/test/java/blackjack/model/participant/PlayerTest.java
@@ -1,6 +1,5 @@
 package blackjack.model.participant;
 
-import blackjack.model.BetAmount;
 import blackjack.model.card.Card;
 import blackjack.model.card.CardDeck;
 import blackjack.model.result.Result;
